don't like the replacement of outliers with median.
Earlier, the outliers were replaced with the median values. However, a quick glance through some of the data suggests that median is "too appropriate" and doesnt really goes good with inflation. So, replacing them with the values of the fence values may be a good substitution (to capture the price hikes). Upon researching, I got to know that the hike in price of onion in 2015 was because of late showers. So this must be included (hint to use upper fence)