multiplicative vs additive
it looks like the seasonality type is multiplicative. reaching to this conclusion on the basis of visual inspection of the mean and standard deviation after transformation and by looking at the spectrum. Clear seasonal peaks observable from the fourier transform of the auto-correlated function for the case of actual data were missing in the case of transformed data. beginning the analysis of the trendline to extract the overall trend (upward or downward) and to deseasonalize the data for comparison