created linear trend algo.
an algorithm to evaluate the linear trend of the prices has been developed. the algorithm is used to generate a file with coefficients defined for equation (y=mx+c) for all the crops. a column containing r sqaured value is also given. it is interesting to note that the crops that are seasonal in nature i.e. follows the sowing and reaping seasons have huge spread along y=0 axis. this is also evident from weak correlation.
NOTE: y=mx+c assumes x to be in months